requirements.txt
Flask==2.0.1
python-telegram-bot==20.0a2
requests==2.26.0